second hw

+++ routes/web.php
@@ -13,16 +13,27 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/', function () {
+Route::get('/', function() {
     return view('welcome');
 });
 
-Route::get('/posts', [\App\Http\Controllers\PostController::class, 'index']);
-Route::get('/posts/{id}', [\App\Http\Controllers\PostController::class, 'show'])->name('posts.show');
-Route::get("/post/create", [\App\Http\Controllers\PostController::class, 'create'])->name('posts.create');
+Route::get('/posts', [\App\Http\Controllers\PostController::class, 'index'])->name('posts')->middleware('auth');
+Route::get('/posts/{id}', [\App\Http\Controllers\PostController::class, 'show'])->name('posts.show')->middleware('auth');
+Route::get("/post/create", [\App\Http\Controllers\PostController::class, 'create'])->name('posts.create')->middleware('auth');
 
-Route::post("/post/save", [\App\Http\Controllers\PostController::class, 'save'])->name('posts.save');
-//
-Route::get('/posts/{id}/edit', [\App\Http\Controllers\PostController::class,"edit"])->name('posts.edit');
-Route::put("posts/{id}/update", [\App\Http\Controllers\PostController::class, "update"])->name("posts.update");
-Route::delete('/posts/{post}/delete', [\App\Http\Controllers\PostController::class, "delete"])->name("posts.delete");
+Route::post("/post/save", [\App\Http\Controllers\PostController::class, 'save'])->name('posts.save')->middleware('auth');
+
+Route::get('/posts/{id}/edit', [\App\Http\Controllers\PostController::class,"edit"])->name('posts.edit')->middleware('auth');
+Route::put("posts/{id}/update", [\App\Http\Controllers\PostController::class, "update"])->name("posts.update")->middleware('auth');
+Route::delete('/posts/{post}/delete', [\App\Http\Controllers\PostController::class, "delete"])->name("posts.delete")->middleware('auth');
+
+Route::get('/users/login', [\App\Http\Controllers\LoginController::class, 'login'])->name('login');
+Route::post('users/post-login', [\App\Http\Controllers\LoginController::class, 'postLogin'])->name('post.login');
+
+Route::post('/users/logout', [\App\Http\Controllers\LoginController::class, 'logout'])->name('logout');
+
+Route::get('/users/register', [\App\Http\Controllers\LoginController::class, 'register'])->name('register');
+Route::post('users/post-register', [\App\Http\Controllers\LoginController::class, 'postRegister'])->name('post.register');
+
+Route::get('/my-posts', [\App\Http\Controllers\LoginController::class, 'myPosts'])->name('my.posts');
+
